opt: move checkExpr to memo
We now call checkExpr from `MemoizeNormExpr`.

Release note: None
