+++ pkg/sql/opt/memo/check_expr.go
@@ -12,27 +12,26 @@
 // implied. See the License for the specific language governing
 // permissions and limitations under the License.
 
-package norm
+package memo
 
 import (
 	"fmt"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/opt"
-	"github.com/cockroachdb/cockroach/pkg/sql/opt/memo"
 	"github.com/cockroachdb/cockroach/pkg/sql/opt/props"
 )
 
-// checkExpr does sanity checking on an Expr. This code is called from
-// onConstruct in testrace builds (which gives us test/CI coverage but elides
-// this code in regular builds).
+// CheckExpr does sanity checking on an Expr. This code is called in testrace
+// builds (which gives us test/CI coverage but elides this code in regular
+// builds).
 // This function does not assume that the expression has been fully normalized.
-func (f *Factory) checkExpr(ev memo.ExprView) {
+func CheckExpr(ev ExprView) {
 	// Check logical properties.
 	ev.Logical().Verify()
 
 	switch ev.Operator() {
 	case opt.ScanOp:
-		def := ev.Private().(*memo.ScanOpDef)
+		def := ev.Private().(*ScanOpDef)
 		if def.Flags.NoIndexJoin && def.Flags.ForceIndex {
 			panic("NoIndexJoin and ForceIndex set")
 		}
@@ -40,7 +39,7 @@ func (f *Factory) checkExpr(ev memo.ExprView) {
 	case opt.ProjectionsOp:
 		// Check that we aren't passing through columns in projection expressions.
 		n := ev.ChildCount()
-		def := ev.Private().(*memo.ProjectionsOpDef)
+		def := ev.Private().(*ProjectionsOpDef)
 		colList := def.SynthesizedCols
 		if len(colList) != n {
 			panic(fmt.Sprintf("%d projections but %d columns", n, len(colList)))
@@ -85,13 +84,13 @@ func (f *Factory) checkExpr(ev memo.ExprView) {
 		}
 
 	case opt.IndexJoinOp:
-		def := ev.Private().(*memo.IndexJoinDef)
+		def := ev.Private().(*IndexJoinDef)
 		if def.Cols.Empty() {
 			panic(fmt.Sprintf("index join with no columns"))
 		}
 
 	case opt.LookupJoinOp:
-		def := ev.Private().(*memo.LookupJoinDef)
+		def := ev.Private().(*LookupJoinDef)
 		if len(def.KeyCols) == 0 {
 			panic(fmt.Sprintf("lookup join with no key columns"))
 		}
@@ -118,11 +117,11 @@ func (f *Factory) checkExpr(ev memo.ExprView) {
 		}
 	}
 
-	f.checkExprOrdering(ev)
+	checkExprOrdering(ev)
 }
 
 // checkExprOrdering runs checks on orderings stored inside operators.
-func (f *Factory) checkExprOrdering(ev memo.ExprView) {
+func checkExprOrdering(ev ExprView) {
 	// Verify that orderings stored in operators only refer to columns produced by
 	// their input.
 	var ordering *props.OrderingChoice
@@ -130,9 +129,9 @@ func (f *Factory) checkExprOrdering(ev memo.ExprView) {
 	case opt.LimitOp, opt.OffsetOp:
 		ordering = ev.Private().(*props.OrderingChoice)
 	case opt.RowNumberOp:
-		ordering = &ev.Private().(*memo.RowNumberDef).Ordering
+		ordering = &ev.Private().(*RowNumberDef).Ordering
 	case opt.GroupByOp, opt.ScalarGroupByOp, opt.DistinctOnOp:
-		ordering = &ev.Private().(*memo.GroupByDef).Ordering
+		ordering = &ev.Private().(*GroupByDef).Ordering
 	default:
 		return
 	}
